functions/parse-api/requirements.txt
functions-framework==3.*
google-cloud-secretmanager
google-cloud-storage
duckdb==1.1.0
pandas
